ldpd, lib, zebra, topotests: add MPLS interface config from LDP
When LDP is enabled per interface, MPLS is not turned on on each
of those interfaces.

Fix this by letting LDP configure the interface if the 'mpls auto'
mode is configured.

Create an IMSG to send interface config request from ldpe thread
to main thread of ldp.
Create a ZAPI message to set MPLS from ldp to zebra.
Update the MPLS setting.
Add a test that ensures that LDP properly configures the interface.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>