isisd: add the prefix-metric support to flex-algo
Add the prefix-metric (aka. M-Flag) support to flex-algo.

When a FAD with the M-flag is elected, the redistributed prefixes (i.e.
from another protocol or another IS-IS level) are advertised with an
additional specific per algorithm Flex-Algo metric (advertisements were
implemented in the two previous commits). This metric is used instead of
the 'igp' one when computing the paths. If it is available (probably
because of an incorrect Flex-Algo implementation on another system), the
'igp' metric will not be used and the prefix will be dropped.

The criteria to determine a prefix has been redistributed are:

- The prefix is of the IPv6 type and has the external flag directly in
  the IPv6 Reachability TLV
- or the prefix-SID for the specific algorithm has the Re-advertisement
  flag.

Signed-off-by: G. Paul Ziemba <paulz@labn.net>
Signed-off-by: Louis Scalbert <louis.scalbert@6wind.com>