zebra: Ensure startup time will be considered as earlier
When starting FRR both the Route Entry has a uptime value
that is saved for when FRR first learns about this particular
entry.  When zebra is undergoing a graceful restart situation
this makes it especially interesting in that monotime(NULL)
will just return the number of seconds the box has been up.
Imagine a quick startup scenario where both
the zrouter.startup_time and the re->uptime are all the
same and we are undergoing some sort of graceful
restart operation:

2023/03/10 17:03:30.976320 ZEBRA: [GD025-3BRF7] 90.0.0.1/32 re->uptime is 277 and compare time is 277
2023/03/10 17:03:30.976321 ZEBRA: [GD025-3BRF7] 90.0.0.2/32 re->uptime is 277 and compare time is 277
2023/03/10 17:03:30.976322 ZEBRA: [GD025-3BRF7] 90.0.0.3/32 re->uptime is 277 and compare time is 277
2023/03/10 17:03:30.976323 ZEBRA: [GD025-3BRF7] 90.0.0.4/32 re->uptime is 279 and compare time is 277
2023/03/10 17:03:30.976324 ZEBRA: [GD025-3BRF7] 90.0.0.5/32 re->uptime is 279 and compare time is 277
2023/03/10 17:03:30.976324 ZEBRA: [GD025-3BRF7] 90.0.0.6/32 re->uptime is 279 and compare time is 277
2023/03/10 17:03:30.976325 ZEBRA: [GD025-3BRF7] 90.0.0.7/32 re->uptime is 279 and compare time is 277

The 90.0.0.1 - > .3 routes did not get update(d) and as such the re->uptime is the same as the
zrouter.startup_time.  This leaving us in a situation where FRR does not believe the route is `old`.

As a quick workaround just set the re->uptime second one second earlier on startup.

Signed-off-by: Donald Sharp <sharpd@nvidia.com>