bgpd: handle link-local nexthop for vpnv4 prefix with v6 nexthop
As per RFC5549 Section 3,
This document extends the definition of the AFI/SAFI for advertisement of IPv4 NLRI
   and VPN-IPv4 NLRI to extend the set of network-layer protocols to
   which the Next Hop address can belong, to include IPv6 in addition to
   IPv4.
   Specifically, this document allows advertising with [RFC4760] of an
   MP_REACH_NLRI with:
   -  AFI = 1
   -  SAFI = 1, 2, 4, or 128
   -  Length of Next Hop Address = 16 or 32
   -  Next Hop Address = IPv6 address of next hop (potentially followed
      by the link-local IPv6 address of the next hop).  This field is to
      be constructed as per Section 3 of [RFC2545].

Issue:
On a point to point link, where the peers share common subnet,
vpnv4 NLRIs advertised does not carry link-local address in nexthop field when
extended nexthop capability is configured on the peer.

Fix:
Nexthop field to be constructed as per Section 3 of [RFC2545] for
AFI-SAFI <1,128> when extended nexthop capability is configured on the peer.

Signed-off-by: Samanvitha B Bhargav <bsamanvitha@vmware.com>