nhrpd: add timer information on 'show ip nhrp shortcut' command
When creating a shortcut, there is no information on the life cycle
of the entry. Add some more fields in the json part of the command.

> # show ip nhrp shortcut
> [..]
>   "table":[
>     {
>       "type":"dynamic",
>       "prefix":"11.255.255.4\/32",
>       "via":"11.255.255.4",
>       "identity":"south",
>       "holdingTimeSecs":1200,          <-- added
>       "routeInstalled":true,           <-- added
>       "expiring":false,                <-- added
>       "remainingTimeSecs":312          <-- added
>     }

The remaining time can be understood with the expiring flag.
If true, the timer stands for the remaining time before
shortcut is removed. Otherwise, the timer stands for the time
before a renewal attempt may happen (initially set to the 2
thirds of the 'holdingTimeSecs' value.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>