topotests: nhrp_topo1, ignore the 'used' variable of cache entries
The 'used' attribute is dynamic and relies on the state of the
system neighbor entry attached to the protocol address. Periodically,
the value is set to false, when receiving a NUD_PROBE notification
from netlink. Then, NHRP re-programs the link entry and re-sets the
'used ' value to true.

Because that value may be intermittently set to false, it is better to
remove it from the tests.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>