bgpd: add nexthop group support for single path routes
On a transit BGP device that receives one full route from one
ISP, if a failover happen, then BGP needs to remove the
installed routes... which may take time.

This takes time because the ZAPI interface between BGP and ZEBRA,
that is in charge of transmitting routes add/updates/removal has
to process a lot of route messages during the transition time.

To reduce the number of messages at ZAPI level, a series of
patches is proposed:
- by splitting the route information in two, and separating the
prefix information from the nexthop information. Two different
messages will be issued
- by avoiding sending route updates when a only the identified
nexthop has been changed.

This commit relies on the nexthop group library extensions.
- A nexthop group identifier is created at BGP level, and reuses
the RECURSION, IBGP and SRTE flags defined in nexthop_group.h
- The nexthop group NHG_ADD message is sent before ROUTE_ADD is
sent with the nexthop group identifier.
- The NHG_DELETE message is sent when the last ROUTE_DEL message
is sent.

The changes :
- only apply to BGP paths which have a single nexthop.
- do not apply to EVPN routes or directly connected routes or blackhole
routes, or ipv4 mapped ipv6 nexthops
- do not apply to routes resolved over blackhole, default route
- do not apply to non host routes that resolve over themselves

The changes do not avoid sending route updates in this commit.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>