sharpd, lib: add force-nexthop-config option to sharp in NHG
In BGP, it is possible to push an eBGP route without using the
ALLOW_RECURSION flag in the route flags option.

In SHARP, it is possible to do the same with the 'sharp install route'
command, but it is not possible to push a NHG that has not
allow-recursion flag set.

> sharp install routes 9.9.9.20 ... no-recurse
> => pushed
> nexthop-group A
>  nexthop 172.31.0.200
> => not pushed

This is a design issue related to the SHARP daemon only.
Change the design:
- Extend the nexthop group hooks with a running-config call back
- Add a vty command in SHARP only, under the nexthop-group node,
that will permit to push any NHG.

> nexthop-group A
>  force-nexthop-config
>  nexthop 172.31.0.200
>

The nhg will be sent to ZEBRA, and will be attempted to be
resolved.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>