pathd, lib, zebra: srte, add distance and metric values in srv6 segment_list
When notifying an update for a colored nexthop, ZEBRA sends the default PATHD
distance value and the 0 metric value. Those two values break the original
values that originated from an IGP route, for instance. When transmitting this
information to BGP for instance, the distance and values are very important
when computing BGP paths.

Fix this by keeping the original distance and metric values.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>
Signed-off-by: Dmytro Shytyi <dmytro.shytyi@6wind.com>