zebra: Wrap `srv6_locator` in `zebra_srv6_locator`
Currently, we have a struct `srv6_locator` used to store information
related to an SRv6 locator. This struct is located under the `lib/`
folder and is therefore shared by all daemons.

This commit adds a new data structure, `zebra_srv6_locator`, specific to
zebra, and positioned under the `zebra/` folder. This data structure is
simply a wrapper around the existing `srv6_locator`. It allows enriching
the SRv6 locator by adding information specific to zebra, thus avoiding
the need to extend the daemon-independent `srv6_locator` data structure
with details that are only relevant to zebra.

Signed-off-by: Carmine Scarpitta <cscarpit@cisco.com>