bgpd: Request SRv6 locator info
Currently, the SRv6 Manager assigns a chunk from the locator to BGP, and
BGP is responsible for autonomously allocating SIDs from that chunk.

Previous commits have extended the SRv6 SID Manager to support APIs for
allocating/releasing SRv6 SIDs to daemons upon request.

This commit aligns the implementation of BGP with the new SRv6 SID Manager.
Specifically, when the locator is configured in BGP, BGP requests
information about the locator using the GET_SRV6_LOCATOR API exposed by the
SRv6 SID Manager and stores the locator information in its database.

Signed-off-by: Carmine Scarpitta <cscarpit@cisco.com>