bgpd: Request SRv6 SIDs to SID Manager
Currently, the SRv6 Manager assigns a chunk from the locator to BGP,
and BGP is responsible for autonomously allocating SIDs from that
chunk.

Previous commits have extended the SRv6 SID Manager to support APIs for
allocating/releasing SRv6 SIDs to daemons upon request.

This commit modifies BGP to delegate the allocation of SIDs to the SID
Manager. Specifically, BGP uses the GET_SRV6_SID API to obtain an SID
from the SID Manager. Once received, BGP installs the SID in the RIB
and announces it to peers.

Signed-off-by: Carmine Scarpitta <cscarpit@cisco.com>