isisd: Request SRv6 locator info
Currently, the SRv6 Manager assigns a chunk from the locator to IS-IS, and
IS-IS is responsible for autonomously allocating SIDs from that chunk.

Previous commits have extended the SRv6 SID Manager to support APIs for
allocating/releasing SRv6 SIDs to daemons upon request.

This commit aligns the implementation of IS-IS with the new SRv6 SID Manager.
Specifically, when the locator is configured in IS-IS, IS-IS requests
information about the locator using the GET_SRV6_LOCATOR API exposed by the
SRv6 SID Manager.

Signed-off-by: Carmine Scarpitta <cscarpit@cisco.com>